% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities.R
\name{get_chemical_descriptions}
\alias{get_chemical_descriptions}
\title{Access Descriptions of Chemicals in PDB Files}
\usage{
get_chemical_descriptions(chemical_ids, ...)
}
\description{
Access Descriptions of Chemicals in PDB Files
}

