% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{get_pdb_description}
\alias{get_pdb_description}
\title{Access Descriptions of PDB Files}
\usage{
get_pdb_description(file_ids, ...)
}
\arguments{
\item{file_ids}{a character vector of file IDs. If you don't have them to hand,
these can be retrieved with \code{\link{list_pdbs}}.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{get_pdb_description} allows you to
access metadata about a particular PDB file, including what structures it
contains, when it was released, and what resolution the underlying data has.
}
\examples{
#Return metadata about PDB files
pdb_metadata <- get_pdb_description(c("116L","100D","14PS"))
}
\seealso{
\code{\link{get_structure_descriptions}}, which makes use
of file description metadata.
}

