% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{get_pdb_entities}
\alias{get_pdb_entities}
\title{Access Information about Entities Within PDB Files}
\usage{
get_pdb_entities(file_ids, ...)
}
\arguments{
\item{file_ids}{a character vector of file IDs. If you don't have them to hand,
these can be retrieved with \code{\link{list_pdbs}}.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{get_pdb_entites} retrieves data about the
content of particular PDB files, and the structures within them.
For obsolete files, this is largely the ID of the file that replaced
it.
}
\examples{
#Return metadata about the entities within obsolete and non-obsolete PDB files
pdb_metadata <- get_pdb_entities(c("116L","100D","14PS"))
}

