% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{get_pdb_ligands}
\alias{get_pdb_ligands}
\title{Retrieve Ligand Information for Entities Within PDB Files}
\usage{
get_pdb_ligands(file_ids, ...)
}
\arguments{
\item{file_ids}{a character vector of file IDs. If you don't have them to hand,
these can be retrieved with \code{\link{list_pdbs}}.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{get_release_status} retrieves information about the
ligands found in elements in a particular PDB file.
}

