% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities.R
\name{get_structure_descriptions}
\alias{get_structure_descriptions}
\title{Access Descriptions of Structures in PDB Files}
\usage{
get_structure_descriptions(structure_ids, ...)
}
\description{
Access Descriptions of Structures in PDB Files
}

