% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities.R
\name{get_chemical_descriptions}
\alias{get_chemical_descriptions}
\title{Access Descriptions of Chemicals in PDB Files}
\usage{
get_chemical_descriptions(chemical_ids, ...)
}
\arguments{
\item{...}{further arguments to pass to httr's GET.}

\item{a}{character vector of chemical IDs.}
}
\description{
PDB files and their elements contain individual chemicals as well
as protein structured. Metadata about these chemicals can be retrieved using
\code{get_chemical_descriptions}.
}
\examples{
#Retrieve the description of NAG - n-acetyl-d-glucosamine
nag_desc <- get_chemical_descriptions("NAG")
}

