% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{get_release_status}
\alias{get_release_status}
\title{Retrieve Release Statuses for Entities Within PDB Files}
\usage{
get_release_status(file_ids, ...)
}
\arguments{
\item{file_ids}{a character vector of file IDs. If you don't have them to hand,
these can be retrieved with \code{\link{list_pdbs}}.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{get_release_status} retrieves information about whether
the entities within PDB files have been released or not.
}

