% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entities.R
\name{get_structure_descriptions}
\alias{get_structure_descriptions}
\title{Access Descriptions of Structures in PDB Files}
\usage{
get_structure_descriptions(structure_ids, ...)
}
\arguments{
\item{...}{further arguments to pass to httr's GET.}

\item{a}{character vector of structure IDs. These can be retrieved from
file metadata or lists (see \code{\link{list_pdbs}}.)}
}
\description{
Access Descriptions of Structures in PDB Files
}

