% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{list_pdbs}
\alias{list_pdbs}
\title{List Accessible PDB Files}
\usage{
list_pdbs(type = c("current", "obsolete", "unreleased"), ...)
}
\arguments{
\item{type}{whether to retrieve current, obsolete or unreleased PDB files.
"current" by default.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{list_pdbs} returns a list of accessible PDB files that
can then be passed into other functions within \code{protein}. You can filter
to either current or obsolete files.
}
\examples{
\dontrun{

#Get current files
available_pdb_files <- list_pdbs()

#Get obsolete files
available_obsolete_files <- list_pdbs("obsolete")

}
}

